requirements.txt
langchain==0.0.245
chromadb==0.3.22
openai==0.27.7
pygpt4all==1.1.0
pygptj==2.0.3
llama-cpp-python==0.1.64
urllib3==2.0.2
SQLAlchemy==2.0.15
python-dotenv==1.0.0
pydantic==1.10.9
accelerate==0.20.3
duckdb==0.8.0
guidance==0.0.61
protobuf==3.20.3
nvitop==1.1.2
tqdm==4.65.0
docker==6.1.3
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.15
selenium==4.10.0
lxml==4.9.2
pandas==2.0.3
loralib==0.1.1
scikit-learn==1.2.2
evaluate==0.4.0
datasets==2.14.0
opencv-python==4.7.0.72
pillow==9.5.0
transformers==4.30.2
bitsandbytes==0.39.1
xformers==0.0.20
git+https://github.com/huggingface/peft@03eb378eb914fbee709ff7c86ba5b1d033b89524
https://github.com/PanQiWei/AutoGPTQ/releases/download/v0.2.2/auto_gptq-0.2.2+cu117-cp310-cp310-linux_x86_64.whl
https://github.com/jllllll/exllama/releases/download/0.0.3/exllama-0.0.3+cu117-cp310-cp310-linux_x86_64.whl
gTTS==2.3.2
playsound==1.3.0
pyttsx3==2.90
pyttsx4==3.0.15
streamlit==1.25.0
fastapi==0.99.1
uvicorn==0.23.1
flask==2.3.2
toml==0.10.2
SQLAlchemy==2.0.15